move _xx to xx; fix lint
